plop

+++ Lead-It/Assets/Script/CountryElem.cs
@@ -11,13 +11,13 @@ public class CountryElem : MonoBehaviour {
 	}
 
 	public ElemType	Etype;
-	public int		SliderVal;
-	public int		Popularity;
-	public int		ProfitBonus;
-	public int		ExpenseMalus;
-	public int		ProfitPercent;
-	public int		ExpensePercent;
-	public int		Budget;
+	public decimal		SliderVal;
+	public decimal		Popularity;
+	public decimal		ProfitBonus;
+	public decimal		ExpenseMalus;
+	public decimal		ProfitPercent;
+	public decimal		ExpensePercent;
+	public decimal		Budget;
 
 
 	// Use this for initialization
